Supid Next.js won't let you look at the url params if there isn't a fallback.... Which then has to also be in a strictly client-side component, not page. Why.
